+++ test/mask/default.js
@@ -1,10 +1,10 @@
 import makeTestSuite from '@zoroaster/mask'
 import Context from '../context'
-import myNewPackage from '../../src'
+import compress from '../../src'
 
 export default makeTestSuite('test/result/default', {
   async getResults() {
-    const res = await myNewPackage({
+    const res = await compress({
       text: this.input,
     })
     return res