+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-## {{ create_date }}
+## 23 December 2019
 
 ### 0.0.0-pre
 
-- Create `{{ package-name }}` with _[`My New Package`](https://MNPJS.org)_
+- Create `@goa/compress` with _[`My New Package`](https://MNPJS.org)_
 - [repository]: `src`, [`test`](https://contexttesting.com), [`documentary`](https://readme.page) & [`types`](https://typedef.page).